auth_service/requirements.txt
Django==5.2.1
djangorestframework==3.15.1
PyJWT==2.8.0
requests==2.31.0
bcrypt==4.1.2
django-cors-headers==4.3.1
coverage==7.3.2